Special Commit for testing Kawa
This commit updates the companion so it can accept Scheme forms over
HTTP and execute them directly. Because the code needs to be URL
encoded, I have included a “driver” program at the top of the source
tree.

Start the companion on your device taking note of its IP
address. Enter any six-character code and press “Connect with
code” (this turns the network layer on). Then on your computer invoke
the driver:

   python driver.py <ipaddr of phone>

This will give a simple REPL loop. Enter “quit” to leave.

Enjoy.

Change-Id: Ie018567d31deeac571cbd23c974581fd9ae5c27f