Do not reinitialize firebase if project bucket not changed
When setting the project bucket, see if it is being set to the same
value and if so don’t re-initialize the firebase connection. Doing so
had the side effect of triggering DataChanged events of all variables,
which is not desirable if we are not actually changing the project
bucket.

Change-Id: Ia0ba736c96e3da38da05ad4143c41ce800444ad7